requirements.txt
streamlit==0.87.0
Pillow==8.2.0
numpy==1.21.0
tensorflow==2.5.0
opencv-python-headless==4.5.2.54
